캡처 UI 및 기능 수정, 예외 처리 등
- AEP 프로그램 미실행, 풀스크린 미설정 시 예외 처리
- 네트워크 서버 연결 불가능시 예외 처리
- 네트워크 전송 response fail일 경우 예외 처리
- 해상도 UI 추가
- 해상도 및 캡처 단위 TextBox 숫자만 입력되도록 제한
- 캡쳐 클래스 메서드 분리
- 로컬 리소스 삭제 소스 수정

+++ PeakDetector/DetectiveProcess/Resource.cs
@@ -41,12 +41,15 @@ namespace PeakDetector.DetectiveProcess
 
         public void deleteLocalResource(ListView listViewRes)
         {
-            foreach (ListViewItem item in listViewRes.CheckedItems)
+            if(listViewRes.CheckedItems.Count >0)
             {
-                String fileName = item.SubItems[1].Text;
-                File.Delete(FILE_PATH + "\\" + fileName);
-            }
-            this.loadLocalResource(listViewRes);
+                foreach (ListViewItem item in listViewRes.CheckedItems)
+                {
+                    String fileName = item.SubItems[1].Text;
+                    File.Delete(FILE_PATH + "\\" + fileName);
+                }
+                this.loadLocalResource(listViewRes);
+            }   
         }
     }
 }
